xlib: support color glyphs via fallback
When we see color glyphs, bail out of the composite_glyphs function
in the traps compositor. Since the mask compositor always uses
an A8 mask for glyphs, bypass it completely for glyphs.

This is not the most optimal solution, but it works. Possible
improvements here would be to identify fonts that contain any
color glyphs, and only bypass the mask compositor for such fonts.