Support color glyphs in the xlib compositor
Since RENDER extension does not have support for using glyphs as
source, we bail out of the XRender-based implementation of
composite_glyphs when we find any color glyphs. The fallback
leads us to the mask compositor, whose glyphs() implementation
always uses an A8 maks for glyphs, so we have to bypass this
as well.

Since we don't have access to the glyphs loaded by the previous
compositor, we use the has_color_glyphs information of the font
to determine whether we can use the mask compositor.